backend/ai-engine/requirements.txt
# FastAPI + Uvicorn
fastapi>=0.95.0
uvicorn[standard]>=0.21.1

# Пайплайн конфигураций и валидации
pydantic>=2.0
PyYAML>=6.0

# HTTP-клиент для вызовов внешних API
httpx>=0.24

# OpenAI + LangChain
openai>=0.27.0
langchain>=0.0.340
langchain-community>=0.1      # обязательно для новых импортов

# Работа с БД и кэшем
pymongo>=4.0
redis>=4.0

# лемматизация
pymorphy2==0.9.1
# ставим словари из диапазона 2.4.x, чтобы не привязываться к конкретной нестабильной сборке
pymorphy2-dicts-ru>=2.4.403921.4377886,<2.5

# чтобы в контейнере был pkg_resources
setuptools>=60.0
